requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-dotenv==1.0.1
mysqlclient==2.2.4
Jinja2==3.1.4
email-validator==2.2.0
python-multipart==0.0.9   # ← 추가

